Move NoOpRecycleBin
