+++ hedera-state-validator/src/main/java/com/hedera/statevalidation/util/PlatformContextHelper.java
@@ -9,7 +9,7 @@ import org.hiero.base.concurrent.ExecutorFactory;
 import org.hiero.base.file.FileSystemManager;
 import org.hiero.consensus.PathsConfig;
 import org.hiero.consensus.fakes.noop.NoOpMetrics;
-import org.hiero.consensus.io.NoOpRecycleBin;
+import org.hiero.consensus.fakes.noop.NoOpRecycleBin;
 import org.hiero.consensus.io.RecycleBin;
 
 /**
