+++ platform-sdk/consensus-fakes/src/main/java/module-info.java
@@ -6,5 +6,6 @@ module org.hiero.consensus.fakes {
     requires transitive org.hiero.consensus.metrics;
     requires transitive org.hiero.consensus.model;
     requires transitive org.hiero.consensus.roster;
+    requires transitive org.hiero.consensus.utility;
     requires static transitive com.github.spotbugs.annotations;
 }
