+++ platform-sdk/consensus-fakes/src/main/java/org/hiero/consensus/fakes/noop/NoOpRecycleBin.java
@@ -1,9 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.consensus.io;
+package org.hiero.consensus.fakes.noop;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.io.IOException;
 import java.nio.file.Path;
+import org.hiero.consensus.io.RecycleBin;
 
 /**
  * A no-op {@link RecycleBin} implementation.
