+++ platform-sdk/consensus-gui/src/test/java/org/hiero/consensus/gui/HashgraphGuiFromPcesMain.java
@@ -12,9 +12,9 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.nio.file.Path;
 import org.hiero.consensus.fakes.noop.NoOpMetrics;
+import org.hiero.consensus.fakes.noop.NoOpRecycleBin;
 import org.hiero.consensus.gui.api.TestGuiSource;
 import org.hiero.consensus.hashgraph.config.ConsensusConfig;
-import org.hiero.consensus.io.NoOpRecycleBin;
 import org.hiero.consensus.io.RecycleBin;
 import org.hiero.consensus.model.hashgraph.GenesisSnapshotFactory;
 import org.hiero.consensus.pcli.graph.PcesEventGraphSource;
