+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/module-info.java
@@ -79,6 +79,7 @@ module org.hiero.otter.fixtures {
     requires org.hiero.base.concurrent;
     requires org.hiero.consensus.concurrent;
     requires org.hiero.consensus.event.stream;
+    requires org.hiero.consensus.fakes;
     requires org.hiero.consensus.hashgraph.impl.test.fixtures;
     requires org.hiero.consensus.hashgraph;
     requires org.hiero.consensus.iss.detection;
