+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/internal/result/SingleNodePcesResultImpl.java
@@ -13,8 +13,8 @@ import java.nio.file.Path;
 import java.util.Iterator;
 import org.hiero.base.file.FileSystemManager;
 import org.hiero.consensus.PathsConfig;
+import org.hiero.consensus.fakes.noop.NoOpRecycleBin;
 import org.hiero.consensus.io.IOIterator;
-import org.hiero.consensus.io.NoOpRecycleBin;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.node.NodeId;
 import org.hiero.consensus.pces.impl.common.PcesFile;
