+++ platform-sdk/swirlds-cli/src/main/java/org/hiero/consensus/pcli/graph/PcesGraphSlicer.java
@@ -25,8 +25,8 @@ import org.hiero.base.crypto.Signature;
 import org.hiero.consensus.crypto.PbjStreamHasher;
 import org.hiero.consensus.crypto.PlatformSigner;
 import org.hiero.consensus.fakes.noop.NoOpMetrics;
+import org.hiero.consensus.fakes.noop.NoOpRecycleBin;
 import org.hiero.consensus.hashgraph.config.ConsensusConfig;
-import org.hiero.consensus.io.NoOpRecycleBin;
 import org.hiero.consensus.model.event.EventDescriptorWrapper;
 import org.hiero.consensus.model.event.EventOrigin;
 import org.hiero.consensus.model.event.PlatformEvent;
